+++ src/metrics/spanner-metrics-exporter.ts
@@ -28,18 +28,15 @@ export const MAX_BATCH_EXPORT_SIZE = 200;
  * Format and sends metrics information to Google Cloud Monitoring.
  */
 export class CloudMonitoringMetricsExporter implements PushMetricExporter {
-  private _projectId: string | void | Promise<string | void>;
+  private _projectId: string;
   private _lastExported: Date = new Date(0);
   private readonly _client: MetricServiceClient;
+  private _metricsExportFailureLogged = false;
 
-  constructor({auth}: ExporterOptions) {
+  constructor({auth}: ExporterOptions, projectId: string) {
     this._client = new MetricServiceClient({auth: auth});
 
-    // Start this async process as early as possible. It will be
-    // awaited on the first export because constructors are synchronous
-    this._projectId = auth.getProjectId().catch(err => {
-      console.error(err);
-    });
+    this._projectId = projectId;
   }
 
   /**
@@ -65,7 +62,6 @@ export class CloudMonitoringMetricsExporter implements PushMetricExporter {
 
     this._lastExported = now;
     this._exportAsync(metrics).then(resultCallback, err => {
-      console.error(err.message);
       resultCallback({code: ExportResultCode.FAILED, error: err});
     });
   }
@@ -83,18 +79,10 @@ export class CloudMonitoringMetricsExporter implements PushMetricExporter {
   private async _exportAsync(
     resourceMetrics: ResourceMetrics,
   ): Promise<ExportResult> {
-    if (this._projectId instanceof Promise) {
-      this._projectId = await this._projectId;
-    }
-
-    if (!this._projectId) {
-      const error = new Error('expecting a non-blank ProjectID');
-      console.error(error.message);
-      return {code: ExportResultCode.FAILED, error};
-    }
-
-    const timeSeriesList =
-      transformResourceMetricToTimeSeriesArray(resourceMetrics);
+    const timeSeriesList = transformResourceMetricToTimeSeriesArray(
+      resourceMetrics,
+      this._projectId,
+    );
 
     let failure: {sendFailed: false} | {sendFailed: true; error: Error} = {
       sendFailed: false,
@@ -104,17 +92,21 @@ export class CloudMonitoringMetricsExporter implements PushMetricExporter {
         async batchedTimeSeries => this._sendTimeSeries(batchedTimeSeries),
       ),
     ).catch(e => {
-      const error = e as {code: number};
-      if (error.code === status.PERMISSION_DENIED) {
-        console.warn(
-          `Need monitoring metric writer permission on project ${this._projectId}. Follow https://cloud.google.com/spanner/docs/view-manage-client-side-metrics#access-client-side-metrics to set up permissions`,
-        );
+      if (!this._metricsExportFailureLogged) {
+        const error = e as {code: number};
+        let msg = 'Send TimeSeries failed:';
+        if (error.code === status.PERMISSION_DENIED) {
+          msg += ` Need monitoring metric writer permission on project ${this._projectId}. Follow https://cloud.google.com/spanner/docs/view-manage-client-side-metrics#access-client-side-metrics to set up permissions`;
+        }
+        console.warn(msg);
+        this._metricsExportFailureLogged = true;
       }
-      const err = asError(e);
-      err.message = `Send TimeSeries failed: ${err.message}`;
-      failure = {sendFailed: true, error: err};
-      console.error(`ERROR: ${err.message}`);
+      failure = {sendFailed: true, error: asError(e)};
     });
+
+    if (!failure.sendFailed && this._metricsExportFailureLogged) {
+      this._metricsExportFailureLogged = false;
+    }
 
     return failure.sendFailed
       ? {